Split hex_utilities.c off from hex.c
The intent is to put all the fiddly hexagonal geometry code in one
place, so that altering the XY or H values doesn't have a slew of
side effects from the dozens of misc. functions they are used in.
Most of those functions are now in hex_utilities; there are a few
more in the various *_draw.c files.