Oh god what even is this
This commit contains: the first worldgen system, which will be replaced, a
mostly-useless piece of UI code, a few changes to the system code and main
loop, a renaming of part of the entity code that affects like every
component, an alteration of the hex utility code that stores radii
internally, a shuffling around of includes, a new vector function, and the
fragments of the /second/ worldgen system.

It's a mess.

But it compiles! D: