Making another mockery of version control here.
This is all the still in-progress changes of the map system; the program
compiles and runs but it's not possible to generate or render the world,
and there's a constant stream of error messages because of that.

There are a few less code files as of this update; that's because certain
*_draw.(c|h) files have been merged with their component_*.(c|h) files,
and (more importantly) the old map system (consisting of component_ground,
ground_draw, world_position, hex, and hex_draw) has been entirely
expurgated. They have been replaced by map.c, which handles all the map
code. This has vastly simplified some other code, specifically parts of
the position and camera component, and so even though it doesn't work yet
I think on the whole this is already an improvement.