Added a new position type, hexPos.
There are four different position structs and here I am adding a
fifth. Ideally this will replace three of them, so that ultimately
there are just subhexes and hexposes. It's intended to be a universal
position, for use with concrete in-game objects and abstract arches
both. Currently it's not of much use, but it'll get some use once I
start writing the worldArch component.