Reversed the array order of the hexPos span code to make it fit with all the rest of the span code
It stored poles at [0] and hexes at [n], when everything else stored
poles at [n] and hexes at [0]. It's mostly an academic difference
except the lack of consistency was confusing me every time I had to
write code dealing with its arrays.

Somewhere in the process of doing this I also fixed the huge bug in
map_posRecalcPlatters that resulted in it almost-always recalculating
to the pole level and failing to resolve anything lower down. It works
correctly now.