Made material code more useful and started using it.
It doesn't make sense for the material list to be in the worldgen
code; ideally hardcoded materials should be loaded from a config file
and softcoded materials (e.g., irradiated materials) should be
generated from those. But for the most part the materials should be
pretty opaque to external fiddling, so no more generating them as
part of the worldgen operation.