Rewrote the input system
The old system was a mess. Ostensibly it supported key selectors that
were multiple keys (e.g., Shift+Up) but in practice it did not and
attempting that would turn the system into a nightmare. This new
system only supports keybinds that are a single key but it does it
much better (and with fewer lines if you excuse the huge
string -> code mapping function)