Took some player-specific UI-interaction out of the input code and put it its own component
This fixes a few issues: one is that previously it was possible to
crash the game by trying to open the map or turn on debug mode in the
menu, since those require the player entity to exist. Another is that
it gives a decent place for player-specific UI commands to go.