The "check" module is now an optional dependency, disabled by default (use ./configure --enable-check to require it)
It wasn't really reasonable to require it as a mandatory dependency, given that it was only used for the unit tests.

Two things to note: even without check, you can still attempt to make check, and that will fail at various stages depending on if you actually have check or not; and the unit tests don't actually compile, since they are /egregiously/ outdated and i haven't been using them at all. gj.